starbucks-clone-flask/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Migrate
Flask-Login
gunicorn
python-dotenv